pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "dune-queries"
version = "0.1.0"
description = "Dune Analytics query management and smoke testing tools"
readme = "README.md"
license = {text = "Apache-2.0"}
requires-python = ">=3.11"
dependencies = [
    "dune-client>=1.7.0",
    "python-dotenv>=1.0.0",
    "pandas>=2.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
]

[project.scripts]
smoke-runner = "scripts.smoke_runner:main"
registry-manager = "scripts.registry_manager:main"

[tool.setuptools.packages.find]
include = ["scripts*"]
